1、技术文档中增加对一些常用命令的说明 2、启动服务的Shell脚本增加对JAVA_HOME环境变量的判断，并输出JAVA_HOME

+++ zywork-ucenter/zywork-ucenter-service-impl/src/main/resources/zywork-ucenter-service.sh
@@ -15,44 +15,49 @@ PID_FILE=$SERVICE_NAME\.pid
 
 cd $SERVICE_DIR
 
-case "$1" in
+if [ "$JAVA_HOME" == "" ]; then
+  echo "JAVA_HOME is not set! Please add it..."
+else
+  echo "use JAVA_HOME:$JAVA_HOME"
+  case "$1" in
 
-    start)
-        nohup $JAVA_HOME/java -Xms256m -Xmx512m -Ddubbo.spring.config=classpath*:/config/*.xml -jar $JAR_NAME > /dev/null 2>&1 &
-        echo "the pid is:$!"
-        echo $! > $SERVICE_DIR/$PID_FILE
-        echo "start $SERVICE_NAME"
-        ;;
+      start)
+          nohup $JAVA_HOME/java -Xms256m -Xmx512m -Ddubbo.spring.config=classpath*:/config/*.xml -jar $JAR_NAME > /dev/null 2>&1 &
+          echo "the pid is:$!"
+          echo $! > $SERVICE_DIR/$PID_FILE
+          echo "start $SERVICE_NAME"
+          ;;
 
-    stop)
-        kill `cat $SERVICE_DIR/$PID_FILE`
-        rm -rf $SERVICE_DIR/$PID_FILE
-        echo "stop $SERVICE_NAME"
+      stop)
+          kill `cat $SERVICE_DIR/$PID_FILE`
+          rm -rf $SERVICE_DIR/$PID_FILE
+          echo "stop $SERVICE_NAME"
 
-        sleep 5
+          sleep 5
 
-        PID=`ps -ef | grep -w "$SERVICE_NAME" | grep 'java' | grep -v "grep" | awk '{print $2}'`
-        if [ "$PID" == "" ]; then
-            echo "$SERVICE_NAME process not exists or stop successfully"
-        else
-            echo "$SERVICE_NAME process pid is:$PID"
-            echo "begin kill $SERVICE_NAME process forcelly, pid is:$PID"
-            kill -9 $PID
-        fi
-        ;;
+          PID=`ps -ef | grep -w "$SERVICE_NAME" | grep 'java' | grep -v "grep" | awk '{print $2}'`
+          if [ "$PID" == "" ]; then
+              echo "$SERVICE_NAME process not exists or stop successfully"
+          else
+              echo "$SERVICE_NAME process pid is:$PID"
+              echo "begin kill $SERVICE_NAME process forcelly, pid is:$PID"
+              kill -9 $PID
+          fi
+          ;;
 
-    restart)
-        $0 stop
-        sleep 2
-        $0 start
-        echo "restarted $SERVICE_NAME"
-        ;;
+      restart)
+          $0 stop
+          sleep 2
+          $0 start
+          echo "restarted $SERVICE_NAME"
+          ;;
 
-    *)
-        $0 stop
-        sleep 2
-        $0 start
-        ;;
+      *)
+          $0 stop
+          sleep 2
+          $0 start
+          ;;
 
-esac
+  esac
+fi
 exit 0
